Add append-only parameter

+++ plugins/operators/load_dimension.py
@@ -3,7 +3,9 @@ from airflow.models import BaseOperator
 from airflow.utils.decorators import apply_defaults
 
 class LoadDimensionOperator(BaseOperator):
-
+    """ Operator for creating and loading dimension tables
+    """
+    
     ui_color = '#80BD9E'
 
     @apply_defaults
@@ -15,6 +17,7 @@ class LoadDimensionOperator(BaseOperator):
                  table="",
                  create_table_sql="",
                  table_insert_sql="",
+                 append_only=False,
                  *args, **kwargs):
 
         super(LoadDimensionOperator, self).__init__(*args, **kwargs)
@@ -25,14 +28,16 @@ class LoadDimensionOperator(BaseOperator):
         self.table_insert_sql=table_insert_sql
         self.redshift_conn_id=redshift_conn_id
         self.table=table
-        
+        self.append_only=append_only
 
     def execute(self, context):
         self.log.info('LoadDimensionOperator implemented')
         redshift = PostgresHook(postgres_conn_id=self.redshift_conn_id)
-        self.log.info(f"Drop then create table: {self.table}")
-        redshift.run(f"DROP TABLE IF EXISTS public.{self.table}")
-        redshift.run(self.create_table_sql)
+        
+        if not self.append_only:
+            self.log.info(f"Drop then create table: {self.table}")
+            redshift.run(f"DROP TABLE IF EXISTS public.{self.table}")
+            redshift.run(self.create_table_sql)
 
         self.log.info(f"Instert data into table: {self.table}")
         redshift.run(self.table_insert_sql)
